+++ src/strings.rs
@@ -141,7 +141,6 @@ impl Display for ProductionString {
 }
 
 
-
 #[cfg(test)]
 mod tests {
     use super::*;